565 - dsl - Add wrappers
