+++ server/sdks/java/component-api/src/main/java/com/bytechef/component/definition/ActionDefinition.java
@@ -122,5 +122,4 @@ public interface ActionDefinition {
         Object apply(Parameters inputParameters, Parameters connectionParameters, ActionContext context)
             throws Exception;
     }
-
 }
